Bug 1516458 - improve handling of env vars
* factor out and test non-platform-specific MergeEnvLists
* handle the task-user and current-user modes differently
